# openapi.yml
openapi: 3.0.3
info:
  title: AssetFlow - Organisation Setup API
  description: >-
    Admin-facing Organisation Setup module of the AssetFlow ERP, part of
    java-main-backend. Manages departments, asset categories, the employee
    directory, organisation settings, and an org-setup dashboard. All endpoints
    are served under the base path /api/v1/admin and require a bearer JWT issued
    by the authentication module.
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development

security:
  - bearerAuth: []

tags:
  - name: Departments
  - name: Asset Categories
  - name: Employees
  - name: Organisation
  - name: Dashboard

paths:
  /api/v1/admin/departments:
    get:
      tags: [Departments]
      summary: List departments
      parameters:
        - $ref: '#/components/parameters/Search'
        - name: active
          in: query
          schema: { type: boolean }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: sortBy
          in: query
          schema: { type: string, enum: [departmentName, createdAt, updatedAt] }
        - $ref: '#/components/parameters/Direction'
      responses:
        '200':
          description: Paged list of departments
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepartmentPageResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Departments]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDepartmentRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepartmentResponseEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }

  /api/v1/admin/departments/{id}:
    parameters:
      - $ref: '#/components/parameters/PathId'
    get:
      tags: [Departments]
      summary: Get department detail
      responses:
        '200':
          description: Department detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepartmentDetailEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Departments]
      summary: Update a department
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateDepartmentRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepartmentResponseEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

  /api/v1/admin/departments/{id}/activate:
    parameters: [ { $ref: '#/components/parameters/PathId' } ]
    patch:
      tags: [Departments]
      summary: Activate a department
      responses:
        '200':
          description: Activated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepartmentResponseEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }

  /api/v1/admin/departments/{id}/deactivate:
    parameters: [ { $ref: '#/components/parameters/PathId' } ]
    patch:
      tags: [Departments]
      summary: Deactivate a department
      description: Rejected while active employees remain assigned to the department.
      responses:
        '200':
          description: Deactivated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepartmentResponseEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

  /api/v1/admin/departments/{id}/head:
    parameters: [ { $ref: '#/components/parameters/PathId' } ]
    put:
      tags: [Departments]
      summary: Assign department head
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssignHeadRequest' }
      responses:
        '200':
          description: Head assigned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepartmentResponseEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Departments]
      summary: Remove department head
      responses:
        '200':
          description: Head removed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DepartmentResponseEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }

  /api/v1/admin/asset-categories:
    get:
      tags: [Asset Categories]
      summary: List asset categories
      parameters:
        - $ref: '#/components/parameters/Search'
        - name: active
          in: query
          schema: { type: boolean }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: sortBy
          in: query
          schema: { type: string, enum: [categoryName, createdAt, updatedAt] }
        - $ref: '#/components/parameters/Direction'
      responses:
        '200':
          description: Paged list of categories
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryPageResponse' }
    post:
      tags: [Asset Categories]
      summary: Create an asset category
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCategoryRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryResponseEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }

  /api/v1/admin/asset-categories/{id}:
    parameters: [ { $ref: '#/components/parameters/PathId' } ]
    get:
      tags: [Asset Categories]
      summary: Get category detail
      responses:
        '200':
          description: Category
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryResponseEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Asset Categories]
      summary: Update a category
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateCategoryRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryResponseEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

  /api/v1/admin/asset-categories/{id}/activate:
    parameters: [ { $ref: '#/components/parameters/PathId' } ]
    patch:
      tags: [Asset Categories]
      summary: Activate a category
      responses:
        '200':
          description: Activated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryResponseEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }

  /api/v1/admin/asset-categories/{id}/deactivate:
    parameters: [ { $ref: '#/components/parameters/PathId' } ]
    patch:
      tags: [Asset Categories]
      summary: Deactivate a category
      responses:
        '200':
          description: Deactivated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryResponseEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }

  /api/v1/admin/employees:
    get:
      tags: [Employees]
      summary: List employees
      parameters:
        - name: search
          in: query
          description: Matches full name, email, or employee code
          schema: { type: string }
        - name: role
          in: query
          schema: { $ref: '#/components/schemas/Role' }
        - name: departmentId
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/AccountStatus' }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: sortBy
          in: query
          schema: { type: string, enum: [fullName, joiningDate, createdAt] }
        - $ref: '#/components/parameters/Direction'
      responses:
        '200':
          description: Paged list of employees
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmployeePageResponse' }

  /api/v1/admin/employees/statistics:
    get:
      tags: [Employees]
      summary: Employee statistics
      responses:
        '200':
          description: Aggregate counts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmployeeStatisticsEnvelope' }

  /api/v1/admin/employees/me:
    get:
      tags: [Employees]
      summary: Get own profile
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmployeeDetailEnvelope' }
    patch:
      tags: [Employees]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateEmployeeRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmployeeDetailEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }

  /api/v1/admin/employees/{id}:
    parameters: [ { $ref: '#/components/parameters/PathId' } ]
    get:
      tags: [Employees]
      summary: Get employee detail
      responses:
        '200':
          description: Employee detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmployeeDetailEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Employees]
      summary: Admin update of employee profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateEmployeeRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmployeeDetailEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }

  /api/v1/admin/employees/{id}/approve:
    parameters: [ { $ref: '#/components/parameters/PathId' } ]
    patch:
      tags: [Employees]
      summary: Approve a pending employee
      responses:
        '200': { $ref: '#/components/responses/EmployeeDetailOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }

  /api/v1/admin/employees/{id}/reject:
    parameters: [ { $ref: '#/components/parameters/PathId' } ]
    patch:
      tags: [Employees]
      summary: Reject a pending employee
      responses:
        '200': { $ref: '#/components/responses/EmployeeDetailOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }

  /api/v1/admin/employees/{id}/activate:
    parameters: [ { $ref: '#/components/parameters/PathId' } ]
    patch:
      tags: [Employees]
      summary: Activate a suspended or inactive employee
      responses:
        '200': { $ref: '#/components/responses/EmployeeDetailOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }

  /api/v1/admin/employees/{id}/suspend:
    parameters: [ { $ref: '#/components/parameters/PathId' } ]
    patch:
      tags: [Employees]
      summary: Suspend an active employee
      responses:
        '200': { $ref: '#/components/responses/EmployeeDetailOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }

  /api/v1/admin/employees/{id}/unlock:
    parameters: [ { $ref: '#/components/parameters/PathId' } ]
    patch:
      tags: [Employees]
      summary: Unlock a locked employee
      responses:
        '200': { $ref: '#/components/responses/EmployeeDetailOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }

  /api/v1/admin/employees/{id}/role:
    parameters: [ { $ref: '#/components/parameters/PathId' } ]
    patch:
      tags: [Employees]
      summary: Change employee role
      description: The ADMIN role cannot be granted here, and admins cannot change their own role.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangeRoleRequest' }
      responses:
        '200': { $ref: '#/components/responses/EmployeeDetailOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }

  /api/v1/admin/employees/{id}/department:
    parameters: [ { $ref: '#/components/parameters/PathId' } ]
    put:
      tags: [Employees]
      summary: Assign employee to a department
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssignDepartmentRequest' }
      responses:
        '200': { $ref: '#/components/responses/EmployeeDetailOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Employees]
      summary: Remove employee from their department
      responses:
        '200': { $ref: '#/components/responses/EmployeeDetailOk' }
        '404': { $ref: '#/components/responses/NotFound' }

  /api/v1/admin/organization:
    get:
      tags: [Organisation]
      summary: Get organisation settings
      responses:
        '200':
          description: Organisation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationEnvelope' }
    put:
      tags: [Organisation]
      summary: Update organisation settings
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateOrganizationRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }

  /api/v1/admin/dashboard/organization-setup:
    get:
      tags: [Dashboard]
      summary: Organisation-setup KPI summary
      responses:
        '200':
          description: Counts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DashboardEnvelope' }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Search:
      name: search
      in: query
      schema: { type: string }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Direction:
      name: direction
      in: query
      schema: { type: string, enum: [asc, desc], default: asc }

  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    BadRequest:
      description: Invalid request or state transition
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: Uniqueness or state conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    EmployeeDetailOk:
      description: Employee detail
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EmployeeDetailEnvelope' }

  schemas:
    Role:
      type: string
      enum: [ADMIN, ASSET_MANAGER, DEPARTMENT_HEAD, EMPLOYEE]
    AccountStatus:
      type: string
      enum: [PENDING_APPROVAL, ACTIVE, INACTIVE, LOCKED, SUSPENDED, REJECTED]

    ApiEnvelope:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        timestamp: { type: string, format: date-time }
    ApiError:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: object
              nullable: true
              additionalProperties: { type: string }

    PageMeta:
      type: object
      properties:
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        first: { type: boolean }
        last: { type: boolean }

    CreateDepartmentRequest:
      type: object
      required: [departmentName]
      properties:
        departmentName: { type: string, maxLength: 150 }
        departmentCode: { type: string, maxLength: 20, nullable: true }
        description: { type: string, nullable: true }
        parentDepartmentId: { type: string, format: uuid, nullable: true }
        departmentHeadId: { type: string, format: uuid, nullable: true }
    UpdateDepartmentRequest:
      $ref: '#/components/schemas/CreateDepartmentRequest'
    AssignHeadRequest:
      type: object
      required: [userId]
      properties:
        userId: { type: string, format: uuid }
    DepartmentResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        departmentName: { type: string }
        departmentCode: { type: string, nullable: true }
        description: { type: string, nullable: true }
        parentDepartmentId: { type: string, format: uuid, nullable: true }
        parentDepartmentName: { type: string, nullable: true }
        departmentHeadId: { type: string, format: uuid, nullable: true }
        departmentHeadName: { type: string, nullable: true }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    DepartmentDetailResponse:
      allOf:
        - $ref: '#/components/schemas/DepartmentResponse'
        - type: object
          properties:
            employeeCount: { type: integer, format: int64 }
            subDepartmentCount: { type: integer, format: int64 }
    DepartmentResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/DepartmentResponse' }
    DepartmentDetailEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/DepartmentDetailResponse' }
    DepartmentPageResponse:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/PageMeta'
                - type: object
                  properties:
                    content:
                      type: array
                      items: { $ref: '#/components/schemas/DepartmentResponse' }

    CreateCategoryRequest:
      type: object
      required: [categoryName]
      properties:
        categoryName: { type: string, maxLength: 120 }
        categoryCode: { type: string, maxLength: 20, nullable: true }
        description: { type: string, nullable: true }
    UpdateCategoryRequest:
      $ref: '#/components/schemas/CreateCategoryRequest'
    CategoryResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        categoryName: { type: string }
        categoryCode: { type: string, nullable: true }
        description: { type: string, nullable: true }
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CategoryResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/CategoryResponse' }
    CategoryPageResponse:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/PageMeta'
                - type: object
                  properties:
                    content:
                      type: array
                      items: { $ref: '#/components/schemas/CategoryResponse' }

    UpdateEmployeeRequest:
      type: object
      properties:
        fullName: { type: string, maxLength: 120 }
        phoneNumber: { type: string, maxLength: 20, nullable: true }
        profileImageUrl: { type: string, nullable: true }
    ChangeRoleRequest:
      type: object
      required: [role]
      properties:
        role: { $ref: '#/components/schemas/Role' }
    AssignDepartmentRequest:
      type: object
      required: [departmentId]
      properties:
        departmentId: { type: string, format: uuid }
    EmployeeResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        employeeCode: { type: string, nullable: true }
        fullName: { type: string }
        email: { type: string }
        phoneNumber: { type: string, nullable: true }
        profileImageUrl: { type: string, nullable: true }
        role: { $ref: '#/components/schemas/Role' }
        departmentId: { type: string, format: uuid, nullable: true }
        departmentName: { type: string, nullable: true }
        accountStatus: { $ref: '#/components/schemas/AccountStatus' }
        emailVerified: { type: boolean }
        joiningDate: { type: string, format: date, nullable: true }
        createdAt: { type: string, format: date-time }
    EmployeeDetailResponse:
      allOf:
        - $ref: '#/components/schemas/EmployeeResponse'
        - type: object
          properties:
            failedLoginAttempts: { type: integer }
            lastLoginAt: { type: string, format: date-time, nullable: true }
            lockedUntil: { type: string, format: date-time, nullable: true }
            updatedAt: { type: string, format: date-time }
    EmployeeDetailEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/EmployeeDetailResponse' }
    EmployeePageResponse:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/PageMeta'
                - type: object
                  properties:
                    content:
                      type: array
                      items: { $ref: '#/components/schemas/EmployeeResponse' }
    EmployeeStatistics:
      type: object
      properties:
        total: { type: integer, format: int64 }
        pendingApproval: { type: integer, format: int64 }
        active: { type: integer, format: int64 }
        inactive: { type: integer, format: int64 }
        suspended: { type: integer, format: int64 }
        locked: { type: integer, format: int64 }
        assetManagers: { type: integer, format: int64 }
        departmentHeads: { type: integer, format: int64 }
        employees: { type: integer, format: int64 }
    EmployeeStatisticsEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/EmployeeStatistics' }

    UpdateOrganizationRequest:
      type: object
      properties:
        organizationName: { type: string, maxLength: 200 }
        legalName: { type: string, maxLength: 250 }
        description: { type: string }
        logoUrl: { type: string }
        website: { type: string, maxLength: 255 }
        email: { type: string, format: email, maxLength: 255 }
        phone: { type: string, maxLength: 25 }
        addressLine1: { type: string }
        addressLine2: { type: string }
        city: { type: string }
        state: { type: string }
        postalCode: { type: string }
        country: { type: string }
        timezone: { type: string }
        currency: { type: string }
        dateFormat: { type: string }
    OrganizationResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        organizationName: { type: string }
        organizationCode: { type: string }
        legalName: { type: string, nullable: true }
        description: { type: string, nullable: true }
        logoUrl: { type: string, nullable: true }
        website: { type: string, nullable: true }
        email: { type: string }
        phone: { type: string, nullable: true }
        addressLine1: { type: string, nullable: true }
        addressLine2: { type: string, nullable: true }
        city: { type: string, nullable: true }
        state: { type: string, nullable: true }
        postalCode: { type: string, nullable: true }
        country: { type: string, nullable: true }
        timezone: { type: string }
        currency: { type: string }
        dateFormat: { type: string }
        subscriptionPlan: { type: string }
        maxUsers: { type: integer }
        maxAssets: { type: integer }
        active: { type: boolean }
        verified: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    OrganizationEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/OrganizationResponse' }

    OrgSetupDashboard:
      type: object
      properties:
        departments: { type: integer, format: int64 }
        employees: { type: integer, format: int64 }
        assetManagers: { type: integer, format: int64 }
        departmentHeads: { type: integer, format: int64 }
        categories: { type: integer, format: int64 }
        pendingApprovals: { type: integer, format: int64 }
    DashboardEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/OrgSetupDashboard' }